+++ mod.ts
@@ -23,4 +23,10 @@ export { createLuaBridge, runLuaCode, LuaClass, LuaBindings };
 export { LuaBinder, LuaBinding } from './src/lua/bindings.ts';
 export { BridgeError, ErrorCodes } from './src/lua/errors.ts';
 export type { ErrorCode } from './src/lua/errors.ts';
+
+export { default as globalBindings, type GlobalBindings } from './src/bindings/global.ts';
+export { default as jsonBindings, type JsonBindings } from './src/bindings/json.ts';
+export { default as regexBindings, type RegexBindings } from './src/bindings/regex.ts';
+export { default as timersBindings, type TimerBindings } from './src/bindings/timers.ts';
+
 export default LuaBridge;